Se avanza en la instalación y configuración.
Continuar con los casos de prueba.

+++ instalar-go.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 version=$1
 homegoi=$2
+goversion=$3 #parametro opcional.
 
-if [ $EUID -ne 0 ]; then
+#if [ ! $EUID -e 0 ]; then
+if [ ! $(whoami) = "root" ]; then
   echo "Se requieren permisos ROOT para instalar GO"
   echo "Intenta con el comando: sudo sh instalar-go.sh"
   exit 1
@@ -10,8 +12,7 @@ fi
 
 #verificar si existe alguna versión instalada
 
-versioni=$(go version | cut -d " " -f3);
-
+#si existe versión instalada pedir confirmación y eliminarla.
 
 DIR='/usr/local'
 GOROOT="${DIR}/go"
@@ -20,9 +21,34 @@ GOBIN="${GOROOT}/bin"
 # Verificar si existe una versión anterior
 
 if [ -d ${GOROOT} ]; then
-  rm -fr ${GOROOT:-/usr/local/go}
+  if [ $goversion = "" ]; then
+    read -n 2 -p "Desea eliminar la versión instalada? [SI (enter) ó NO]: " confirm;
+  else
+    read -n 2 -p "Desea eliminar la versión: ${goversion}? [SI (enter) ó NO]: " confirm;
+  fi
+  echo $confirm;
+  if [ $confirm = "n" || $confirm = "no" || $confirm = "NO" || $confirm = "N" ]; then
+    echo "Actualización INTERRUMPIDA.";
+    exit 2;
+  else
+    echo "Se reemplazará la versión: go${goversion} por go${version}!";
+    rm -fr ${GOROOT:-/usr/local/go}
+  fi
+
+else
+  echo "NO se encuentra versión de GO instalada!";
+
 fi
 
-tar xzf - -C ${BASE}
+#exit 10;
+
+#Instalar versión indicada.
+archi="${homegoi}/go${version}.linux-amd64.tar.gz";
+echo "Instalando versión: go${version} desde el archivo ${archi} ...";
+tar xzf ${archi} -C ${DIR};
+
+
+# Configurar entorno para Golang
+sh ./configurar-go.sh;
 
 sudo -k
